requirements.txt
torch
torchvision
transformers
timm
gradio
numpy
Pillow